+++ include/matrix.hpp
@@ -3,5 +3,5 @@
 #include <algorithm>
 #include <cstdint>
 
-void reverse_matrix(uint32_t *matrix[], size_t n, size_t m);
-void transpose_matrix(uint32_t *matrix[], size_t n, size_t m);
+void reverse_matrix(uint32_t *matrix[], uint n, uint m);
+void transpose_matrix(uint32_t *matrix[], uint n, uint m);
